feat: 载具预览功能加入

+++ HymsonMES/src/Libraries/Hymson.MES.Services/Services/Integrated/InteVehicle/InteVehicleService.cs
@@ -39,8 +39,9 @@ namespace Hymson.MES.Services.Services.Integrated
 
         private readonly IInteVehicleTypeRepository _inteVehicleTypeRepository;
         private readonly IInteVehicleVerifyRepository _inteVehicleVerifyRepository;
+        private readonly IInteVehicleFreightRepository _inteVehicleFreightRepository;
 
-        public InteVehicleService(ICurrentUser currentUser, ICurrentSite currentSite, IInteVehicleRepository inteVehicleRepository, AbstractValidator<InteVehicleCreateDto> validationCreateRules, AbstractValidator<InteVehicleModifyDto> validationModifyRules,IInteVehicleTypeRepository inteVehicleTypeRepository, IInteVehicleVerifyRepository inteVehicleVerifyRepository)
+        public InteVehicleService(ICurrentUser currentUser, ICurrentSite currentSite, IInteVehicleRepository inteVehicleRepository, AbstractValidator<InteVehicleCreateDto> validationCreateRules, AbstractValidator<InteVehicleModifyDto> validationModifyRules,IInteVehicleTypeRepository inteVehicleTypeRepository, IInteVehicleVerifyRepository inteVehicleVerifyRepository, IInteVehicleFreightRepository inteVehicleFreightRepository)
         {
             _currentUser = currentUser;
             _currentSite = currentSite;
@@ -49,6 +50,7 @@ namespace Hymson.MES.Services.Services.Integrated
             _validationModifyRules = validationModifyRules;
             _inteVehicleTypeRepository = inteVehicleTypeRepository;
             _inteVehicleVerifyRepository = inteVehicleVerifyRepository;
+            _inteVehicleFreightRepository = inteVehicleFreightRepository;
         }
 
         /// <summary>
@@ -108,6 +110,30 @@ namespace Hymson.MES.Services.Services.Integrated
 
             #endregion
 
+            #region 处理 载具装载数据
+            var inteVehicleFreightEntitys = new List<InteVehicleFreightEntity>();
+
+            if (inteVehicleCreateDto.InteVehicleFreights != null && inteVehicleCreateDto.InteVehicleFreights.Any()) 
+            {
+                foreach (var item in inteVehicleCreateDto.InteVehicleFreights)
+                {
+                    inteVehicleFreightEntitys.Add(new InteVehicleFreightEntity()
+                    {
+                        VehicleId = inteVehicleEntity.Id,
+                        Location = item.Location,
+                        Status= item.Status,
+
+                        Id = IdGenProvider.Instance.CreateId(),
+                        CreatedBy = _currentUser.UserName,
+                        UpdatedBy = _currentUser.UserName,
+                        CreatedOn = HymsonClock.Now(),
+                        UpdatedOn = HymsonClock.Now(),
+                        SiteId = _currentSite.SiteId ?? 0
+                    });
+                }
+            }
+            #endregion
+
             using (TransactionScope ts = new TransactionScope())
             {
                 //入库
@@ -116,6 +142,11 @@ namespace Hymson.MES.Services.Services.Integrated
                 if (inteVehicleVerify != null) 
                 {
                     await _inteVehicleVerifyRepository.InsertAsync(inteVehicleVerify);
+                }
+
+                if (inteVehicleFreightEntitys.Any()) 
+                {
+                    await _inteVehicleFreightRepository.InsertsAsync(inteVehicleFreightEntitys);
                 }
 
                 ts.Complete();
@@ -139,6 +170,18 @@ namespace Hymson.MES.Services.Services.Integrated
         /// <returns></returns>
         public async Task<int> DeletesInteVehicleAsync(long[] ids)
         {
+            //校验数据 
+            var inteVehicleFreights= await _inteVehicleFreightRepository.GetByVehicleIdsAsync(ids);
+            var hasBinds = inteVehicleFreights.Where(x => !string.IsNullOrWhiteSpace(x.BarCode));
+            if (hasBinds.Any()) 
+            {
+                //查询是哪些载具
+                var vehicleIds= hasBinds.Select(x => x.VehicleId).Distinct().ToArray();
+                var hasBindVehicles = await _inteVehicleRepository.GetByIdsAsync(vehicleIds);
+
+                throw new CustomerValidationException(nameof(ErrorCode.MES18603)).WithData("codes", string.Join(", ", hasBindVehicles.Select(x => x.Code).Distinct().ToArray()));
+            }
+
             return await _inteVehicleRepository.DeletesAsync(new DeleteCommand { Ids = ids, DeleteOn = HymsonClock.Now(), UserId = _currentUser.UserName });
         }
 
@@ -217,6 +260,31 @@ namespace Hymson.MES.Services.Services.Integrated
 
             #endregion
 
+            #region 处理 载具装载数据
+            var inteVehicleFreightEntitys = new List<InteVehicleFreightEntity>();
+
+            if (inteVehicleModifyDto.InteVehicleFreights != null && inteVehicleModifyDto.InteVehicleFreights.Any())
+            {
+                foreach (var item in inteVehicleModifyDto.InteVehicleFreights)
+                {
+                    inteVehicleFreightEntitys.Add(new InteVehicleFreightEntity()
+                    {
+                        VehicleId = inteVehicleEntity.Id,
+                        Location = item.Location,
+                        Status = item.Status,
+
+                        Id = IdGenProvider.Instance.CreateId(),
+                        CreatedBy = _currentUser.UserName,
+                        UpdatedBy = _currentUser.UserName,
+                        CreatedOn = HymsonClock.Now(),
+                        UpdatedOn = HymsonClock.Now(),
+                        SiteId = _currentSite.SiteId ?? 0
+                    });
+                }
+            }
+            #endregion
+
+
             using (TransactionScope ts = new TransactionScope())
             {
                 await _inteVehicleRepository.UpdateAsync(inteVehicleEntity);
@@ -228,10 +296,14 @@ namespace Hymson.MES.Services.Services.Integrated
                     await _inteVehicleVerifyRepository.InsertAsync(inteVehicleVerify);
                 }
 
+                await _inteVehicleFreightRepository.DeletesTrueByVehicleIdsAsync(new long[] { inteVehicleEntity.Id });
+                if (inteVehicleFreightEntitys.Any())
+                {
+                    await _inteVehicleFreightRepository.InsertsAsync(inteVehicleFreightEntitys);
+                }
+
                 ts.Complete();
             }
-
-            
         }
 
         /// <summary>
@@ -259,7 +331,7 @@ namespace Hymson.MES.Services.Services.Integrated
         }
 
         /// <summary>
-        /// 
+        /// 获取载具验证
         /// </summary>
         /// <param name="vehicleId"></param>
         /// <returns></returns>
@@ -272,6 +344,26 @@ namespace Hymson.MES.Services.Services.Integrated
             }
             else
                 return null;
+        }
+
+        /// <summary>
+        /// 获取载具装载
+        /// </summary>
+        /// <param name="vehicleId"></param>
+        /// <returns></returns>
+        public async Task<IEnumerable<InteVehicleFreightDto>> QueryVehicleFreightByVehicleIdAsync(long vehicleId)
+        {
+            var inteVehicleFreights = await _inteVehicleFreightRepository.GetByVehicleIdsAsync(new long[] { vehicleId });
+            var inteVehicleFreightDtos=new List<InteVehicleFreightDto>();
+            if (inteVehicleFreights != null&& inteVehicleFreights.Any())
+            {
+                foreach (var item in inteVehicleFreights)
+                {
+                    inteVehicleFreightDtos.Add(item.ToModel<InteVehicleFreightDto>());
+                }
+            }
+
+            return inteVehicleFreightDtos;
         }
     }
 }
